nit: cleanups
