+++ shared/components/molecules/FeaturePages/IconTitleDescriptionCard/IconTitleDescriptionCard.types.ts
@@ -1,0 +1,7 @@
+export type IconTitleDescriptionCardProps = {
+  icon: React.ReactNode
+  iconText: string
+  title: string
+  description: string
+  className?: string
+}
