+++ shared/components/molecules/FeaturePages/IconTitleDescriptionCard/IconTitleDescriptionCardGrid.types.ts
@@ -1,0 +1,11 @@
+export type IconTitleDescriptionCardData = {
+  icon: React.ReactNode
+  iconText: string
+  title: string
+  description: string
+}
+
+export type IconTitleDescriptionCardGridProps = {
+  cards: IconTitleDescriptionCardData[]
+  className?: string
+}
