+++ shared/components/molecules/FeaturePages/IconTitleDescriptionCard/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './IconTitleDescriptionCardGrid.view'
